Refactor large indexing helper methods
